+++ src/main/java/com/example/SmartBuildingBackend/controller/RoomController.java
@@ -31,7 +31,7 @@ public class RoomController {
     private RoomService roomService;
 
     @PostMapping("/{floorId}")
-    public ResponseEntity <RoomDto> addRoom( @PathVariable int floorId,@Valid @RequestBody RoomDto roomDto) {
+    public ResponseEntity <RoomDto> addRoom( @PathVariable Long floorId,@Valid @RequestBody RoomDto roomDto) {
         RoomDto newRoom = roomService.addRoom(floorId, roomDto);
         return ResponseEntity.status(HttpStatus.CREATED).body(newRoom);
     }
@@ -43,19 +43,19 @@ public class RoomController {
     }
     
     @GetMapping("/{room_id}")
-    public ResponseEntity<RoomDto> getRoomById(@PathVariable("room_id") int roomId) {
+    public ResponseEntity<RoomDto> getRoomById(@PathVariable("room_id") Long roomId) {
         RoomDto roomDto = roomService.getRoomById(roomId);
         return ResponseEntity.ok(roomDto);
     }
 
     @PutMapping("/{room_id}")
-    public ResponseEntity<RoomDto> updateRoom (@PathVariable("room_id") int roomId,@Valid @RequestBody RoomDto updateRoom) {
+    public ResponseEntity<RoomDto> updateRoom (@PathVariable("room_id") Long roomId,@Valid @RequestBody RoomDto updateRoom) {
         RoomDto updatedRoom = roomService.updateRoom(roomId, updateRoom);
         return ResponseEntity.status(HttpStatus.CREATED).body(updatedRoom);
     }
 
     @DeleteMapping("/{room_id}")
-    public ResponseEntity<String> deleteRoom (@PathVariable("room_id") int roomId) {
+    public ResponseEntity<String> deleteRoom (@PathVariable("room_id") Long roomId) {
         roomService.deleteRoom(roomId);
         return ResponseEntity.ok("Room deleted successfully");
     }
