+++ src/main/java/com/example/SmartBuildingBackend/service/implementation/RoomServiceImplementation.java
@@ -26,13 +26,13 @@ public class RoomServiceImplementation implements RoomService {
         return rooms.stream().map(RoomMapper::mapToRoomDto).collect(Collectors.toList());    
     }
     @Override
-    public RoomDto getRoomById(int roomId) {
+    public RoomDto getRoomById(Long roomId) {
         Room room = roomRepository.findById(roomId)
                 .orElseThrow(() -> new RuntimeException("Room not found with id: " + roomId));
         return RoomMapper.mapToRoomDto(room);
     }
     @Override
-    public RoomDto updateRoom(int roomId, RoomDto updateRoom) {
+    public RoomDto updateRoom(Long roomId, RoomDto updateRoom) {
         Room room = roomRepository.findById(roomId)
                 .orElseThrow(() -> new RuntimeException("Room not found with id: " + roomId));
         
@@ -47,13 +47,13 @@ public class RoomServiceImplementation implements RoomService {
         return RoomMapper.mapToRoomDto(updatedRoom);
     }
     @Override
-    public void deleteRoom(int roomId) {
+    public void deleteRoom(Long roomId) {
         Room room = roomRepository.findById(roomId)
                 .orElseThrow(() -> new RuntimeException("Room not found with id: " + roomId));
         roomRepository.delete(room);
     }
     @Override
-    public RoomDto addRoom(int floorId, RoomDto roomDto) {
+    public RoomDto addRoom(Long floorId, RoomDto roomDto) {
         Floor floor = floorRepository.findById(floorId)
                 .orElseThrow(() -> new RuntimeException("Floor not found with id: " + floorId));
         Room room = new Room();
